backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0

dj-database-url>=2.0,<3.0
psycopg>=3.3.2

requests>=2.31
redis>=7.3.0
gunicorn>=22.0
python-dotenv>=1.0
